Fix #94: Move preferences to mobile-config-autoconfig.js
We only want to keep preferences needed to load our autoconfig file
in the static mobile-config-prefs.js file. For that, we are moving
the extra preferences to our set_default_prefs function in
mobile-config-autoconfig.js. One small caveat, we had to keep the
preferences in the static file because it was not set correctly when
starting Firefox. This will be done in a follow-up MR.

Part-of: https://gitlab.postmarketos.org/postmarketOS/mobile-config-firefox/-/merge_requests/69