autoconfig: add comment with ref to furios repo
The UA agent spoofing code is heavily based on the code from the
FuriOS team (https://github.com/FuriLabs/furios-firefox-tweaks).
This repo in turn is a fork of mobile-config-firefox, under the same
license (MPL-2.0). Add a comment refering to that work to give credit
and because it is probably useful for comparing UA rules in the future.

Part-of: https://gitlab.postmarketos.org/postmarketOS/mobile-config-firefox/-/merge_requests/77